+++ libs/eavmlib/src/esp.erl
@@ -46,7 +46,8 @@
     rtc_slow_get_binary/0,
     rtc_slow_set_binary/1,
     freq_hz/0,
-    get_mac/1
+    get_mac/1,
+    get_default_mac/0
 ]).
 
 -deprecated([
@@ -386,4 +387,18 @@ freq_hz() ->
 %%-----------------------------------------------------------------------------
 -spec get_mac(Interface :: interface()) -> mac().
 get_mac(_Interface) ->
+    erlang:nif_error(undefined).
+
+%%-----------------------------------------------------------------------------
+%% @returns The default MAC address of the ESP32 device.
+%% @doc     Retrieve the default MAC address of the ESP32 device.
+%%          This function accesses the EFUSE memory of the ESP32 and reads
+%%          the factory-programmed MAC address.
+%%
+%%          The mac address is returned as a 6-byte binary, per the
+%%          IEEE 802 family of specifications.
+%% @end
+%%-----------------------------------------------------------------------------
+-spec get_default_mac() -> {ok, mac()} | {error, atom()}.
+get_default_mac() ->
     erlang:nif_error(undefined).